pyproject.toml
[tool.poetry]
name = "savememoney"
version = "0.1.0"
description = ""
authors = ["Udi Oron <udioron@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
Django = "^4.2"
ipython = "^8.12.0"
django-extensions = "^3.2.1"
black = {extras = ["d"], version = "^23.3.0"}
pytest-django = "^4.5.2"
tqdm = "^4.65.0"
faker = "^18.4.0"
django-bootstrap5 = "^23.1"
djangorestframework = "^3.14.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "savememoney.settings"
